added mlp3 and mlp4
These operate on messages directly.  One benefit is that this
correctly avoids 'double counting' in the computation graph
(using the other MLPs, was subtracting pre-MLP messages from
post MLP aggregated message beliefs).  Another benefit is that this
avoids operating on expanded variable to factor messages, which
doesn’t seem to  make sense, because there should be some permutation
invariance to the expansion ordering.